Change the Folder Structure
