+++ frontend/src/components/WorkoutSupportManagement/Workouts/Workout/Workout.js
@@ -1,17 +1,22 @@
 import React from "react";
-import { Grid, ButtonBase } from "@material-ui/core";
+import { Button, Grid, ButtonBase } from "@material-ui/core";
+import DeleteIcon from "@material-ui/icons/Delete";
+import EditIcon from "@material-ui/icons/Edit";
 import moment from "moment";
+import { Link } from "react-router-dom";
 import { useDispatch } from "react-redux";
+import { deleteWorkout } from "../../../../actions/workouts";
 import { useHistory } from "react-router";
 
-const WorkoutDisplayMember = ({ workout, setCurrentId }) => {
+const Workout = ({ workout, setCurrentId }) => {
+  const dispatch = useDispatch();
   const history = useHistory();
   const openWorkout = () => history.push(`/workouts/${workout._id}`);
 
   return (
     <Grid item xs={6} sm={3}>
       <br></br>
-      <div className="card border-dark text-dark">
+      <div class="card border-dark text-dark">
         <div className="card-header">
           {workout.workout_name}
           <br></br>
@@ -45,9 +50,37 @@ const WorkoutDisplayMember = ({ workout, setCurrentId }) => {
             <br></br>
           </div>
         </ButtonBase>
+        <div className="card-header">
+          <Link
+            to={"/trainer/workouts/update/" + workout._id}
+            style={{ textDecoration: "none" }}
+          >
+            <Button
+              size="small"
+              color="primary"
+              onClick={() => setCurrentId(workout._id)}
+            >
+              <EditIcon fontSize="small" />
+              &nbsp; Update
+            </Button>
+          </Link>
+          <Button
+            size="small"
+            color="primary"
+            onClick={() => {
+              if (window.confirm("Are you sure you want to delete this")) {
+                dispatch(deleteWorkout(workout._id));
+                history.push("/trainer/workouts");
+              }
+            }}
+          >
+            <DeleteIcon fontSize="small" />
+            &nbsp; Delete
+          </Button>
+        </div>
       </div>
     </Grid>
   );
 };
 
-export default WorkoutDisplayMember;
+export default Workout;
